requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-core
openai
numpy
faiss-cpu
requests
tiktoken